modify: split holiday from rotation
