+++ src/rotation/entity/rotation-attendee.entity.ts
