+++ src/rotation/entity/rotation.entity.ts
