+++ src/rotation/rotations.repository.ts
@@ -1,10 +1,10 @@
 import { Inject, Injectable, Logger, forwardRef } from '@nestjs/common';
 import { DataSource, Repository } from 'typeorm';
-import { RotationEntity } from './entity/rotation/rotation.entity';
+import { RotationEntity } from './entity/rotation.entity';
 import { getNextYearAndMonth } from './utils/date';
-import { HolidayService } from './holiday.service';
+import { HolidayService } from '../holiday/holiday.service';
 import { RotationsService } from './rotations.service';
-import { RotationAttendeeEntity } from './entity/rotation/rotation-attendee.entity';
+import { RotationAttendeeEntity } from './entity/rotation-attendee.entity';
 import { createRotation } from './utils/rotation';
 import { UserService } from 'src/user/user.service';
 import { CreateRegistrationDto } from './dto/create-registration.dto';
